Restructured project to all depend on shared project.
