+++ IMLLoader/src/Entrypoint.cs
@@ -1,9 +1,4 @@
-﻿using IMLLoader;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using VTOLAPICommons;
 
 namespace Doorstop
 {
