+++ IMLLoader/src/GameStartupPatch.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Reflection;
 using System.Text;
 using UnityEngine;
+using VTOLAPICommons;
 
 namespace IMLLoader
 {
