+++ ModLoaderPolyfill/ModLoaderPolyfill.sln
@@ -1,11 +1,14 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34009.444
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30320.27
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModLoaderPolyfill", "ModLoaderPolyfill.csproj", "{739D5967-86BE-411B-A9D1-6E97349EDCEA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ModLoaderPolyfill", "ModLoaderPolyfill.csproj", "{739D5967-86BE-411B-A9D1-6E97349EDCEA}"
 EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		..\VTOLAPICommons\VTOLAPICommons\VTOLAPICommons.projitems*{739d5967-86be-411b-a9d1-6e97349edcea}*SharedItemsImports = 5
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
