+++ VTOLAPICommons/VTOLAPICommons.sln
@@ -1,0 +1,18 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30320.27
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "VTOLAPICommons", "VTOLAPICommons\VTOLAPICommons.shproj", "{457A7085-D726-496B-B775-9191743CAA4E}"
+EndProject
+Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		VTOLAPICommons\VTOLAPICommons.projitems*{457a7085-d726-496b-b775-9191743caa4e}*SharedItemsImports = 13
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1DF99691-AF9F-4478-A546-5B54939B5AE3}
+	EndGlobalSection
+EndGlobal
