+++ VTOLAPICommons/VTOLAPICommons/Mod/ModInfo.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace IMLLoader
+namespace VTOLAPICommons
 {
     public class ModInfo
     {
