+++ VTOLAPICommons/VTOLAPICommons/WinTrust/TrustProviderFlags.cs
@@ -1,4 +1,4 @@
-﻿namespace IMLLoader.WinTrust
+﻿namespace VTOLAPICommons.WinTrust
 {
     public enum TrustProviderFlags
     {
